Add some constants for creep types determined by the application, and report only battles which contain at least one military-type creep from a player who doesn't own the room.
This should successfully filter out situations where a single hauler or worker wonders into another player's base.

NOTE: this could cause some legitimate battles to be ignore, since it counts creeps with up to 8 work parts as civilians, while creeps with over 8 work parts are dismantlers.

If a small party of 8-work dismantlers attacks a room without any healer or attacker backups, the application will no longer report the battle.
TODO: classify a creep as a dismantler if it uses the 'dismantle' (maybe classified as 'attack' in history view?) action