Move to completely directly using requests to access the Screeps API.
Eventually we should move back to using python-screeps for all API calls, but at the moment it's much easier to handle and log errors specifically by calling the API directly. This way, all errors are correctly logged, and the behavior is as expected. (rather than being printed to console as python-screeps does.)