For now, solve the duplicate battles problem by only scanning every 15 minutes.
In the future, we should ideally be scanning for battles, and keep a queue of *room names* to process, separate from
ticks. We should then keep, for each room name, a list of ticks for which we know hostilities have occurred.

When processing a battle, we should then go through and try to access the history for each known hostilities tick,
and "scan" history backwards, then forwards, until we've found the ends of the combat. If at any point combat still
seems to be continuing when we reach the end of the available history (or there is no available history for a tick
with known hostilities), we should _cache what we know so far_, and leave it in the queue.

The "what we know so far, processing a battle" cache should contain a list of room-name/20-tick-history sections
which we've already retrieved, as well as any accumulative information that we would otherwise be storing in local
variables. The "process battle" function (or set of functions) should be able to pause at and resume paused execution
from any point that we try to get new data.